+++ engine/src/engine/util/mod.rs
@@ -1,4 +1,6 @@
 pub mod buffered_atomic_counter;
+#[cfg(feature = "datagen")]
+pub mod datagen;
 pub mod log;
 pub mod mem;
 pub mod metrics;